requirements.txt
fastapi~=0.91.0
pytz~=2023.3
appwrite~=4.1.0
Telethon~=1.33.1
pymongo~=4.6.1
APScheduler~=3.10.1
starlette~=0.32.0.post1
requests~=2.28.2
Flask~=2.2.5
patool~=2.1.1
FastTelethonhelper~=1.0.7